\name{wbApi-package}
\alias{wbApi-package}
\alias{wbApi}
\docType{package}
\title{
\packageTitle{wbApi}
}
\description{
\packageDescription{wbApi}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{wbApi}
\packageIndices{wbApi}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{wbApi}

Maintainer: \packageMaintainer{wbApi}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
