% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wbapi.R
\name{wbapi}
\alias{wbapi}
\title{wbapi: R Interface to the World Bank API}
\usage{
wbapi()
}
\value{
List of functions that perform calls to the World Bank API
}
\description{
wbapi: R Interface to the World Bank API
}
\author{
Janko Cizel
}

